GH-12 取得したデータを表示する
