--- src/app/leagues/[id]/page.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import { LeagueDetailScreen } from "@/components/screens/league";
-
-export default function LeagueDetailPage({
-  params,
-}: {
-  params: { id: string };
-}) {
-  return <LeagueDetailScreen leagueId={params.id} />;
-}
